Add status collection

+++ src/Infrastructure/Status/Persistence/StatusPersistence.php
@@ -3,18 +3,20 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Status\Persistence;
 
-use App\Aggregate\Repository\TimelyStatusRepository;
 use App\Api\AccessToken\AccessToken;
 use App\Api\Entity\Aggregate;
 use App\Api\Entity\ArchivedStatus;
 use App\Api\Entity\Status;
 use App\Api\Exception\InsertDuplicatesException;
+use App\Domain\Status\StatusCollection;
 use App\Domain\Status\StatusInterface;
 use App\Domain\Status\TaggedStatus;
 use App\Infrastructure\DependencyInjection\StatusLoggerTrait;
 use App\Infrastructure\DependencyInjection\TaggedStatusRepositoryTrait;
+use App\Infrastructure\DependencyInjection\TimelyStatusRepositoryTrait;
+use App\Infrastructure\Repository\Status\TimelyStatusRepositoryInterface;
 use App\Infrastructure\Twitter\Api\Normalizer\Normalizer;
-use DateTime;
+use App\Operation\Collection\CollectionInterface;
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\ORMException;
@@ -26,6 +28,7 @@ class StatusPersistence implements StatusPersistenceInterface
 {
     use StatusLoggerTrait;
     use TaggedStatusRepositoryTrait;
+    use TimelyStatusRepositoryTrait;
 
     public ManagerRegistry $registry;
 
@@ -35,17 +38,12 @@ class StatusPersistence implements StatusPersistenceInterface
     private LoggerInterface $appLogger;
 
     /**
-     * @var TimelyStatusRepository
-     */
-    private TimelyStatusRepository $timelyStatusRepository;
-
-    /**
      * @var EntityManagerInterface
      */
     private EntityManagerInterface $entityManager;
 
     public function __construct(
-        TimelyStatusRepository $timelyStatusRepository,
+        TimelyStatusRepositoryInterface $timelyStatusRepository,
         ManagerRegistry $registry,
         EntityManagerInterface $entityManager,
         LoggerInterface $logger
@@ -63,55 +61,23 @@ class StatusPersistence implements StatusPersistenceInterface
     ): array {
         $propertiesCollection = Normalizer::normalizeAll(
             $statuses,
-            function ($extract) use ($accessToken) {
-                $extract['identifier'] = $accessToken->accessToken();
-
-                return $extract;
-            },
+            $this->tokenSetter($accessToken),
             $this->appLogger
         );
 
-        $statuses = [];
+        $statusCollection = StatusCollection::fromArray([]);
 
         /** @var TaggedStatus $taggedStatus */
         foreach ($propertiesCollection->toArray() as $key => $taggedStatus) {
-            $extract = $taggedStatus->toLegacyProps();
-            $status  = $this->taggedStatusRepository
-                ->convertPropsToStatus($extract, $aggregate);
-
-            if ($status->getId() === null) {
-                $this->collectStatusLogger->logStatus($status);
-            }
-
-            if ($status instanceof ArchivedStatus) {
-                $status = $this->unarchiveStatus($status, $this->entityManager);
-            }
-
             try {
-                if ($status->getId()) {
-                    $status->setUpdatedAt(
-                        new DateTime(
-                            'now',
-                            new \DateTimeZone('UTC')
-                        )
-                    );
-                }
-
-                if ($aggregate instanceof Aggregate) {
-                    $timelyStatus = $this->timelyStatusRepository->fromAggregatedStatus(
-                        $status,
-                        $aggregate
-                    );
-                    $this->entityManager->persist($timelyStatus);
-                }
-
-                $this->entityManager->persist($status);
-
-                $statuses[] = $status;
+                $statusCollection = $this->persistStatus(
+                    $statusCollection,
+                    $taggedStatus,
+                    $aggregate
+                );
             } catch (ORMException $exception) {
                 if ($exception->getMessage() === ORMException::entityManagerClosed()->getMessage()) {
                     $this->entityManager = $this->registry->resetManager('default');
-                    $this->entityManager->persist($status);
                 }
             } catch (Exception $exception) {
                 $this->appLogger->info($exception->getMessage());
@@ -120,10 +86,13 @@ class StatusPersistence implements StatusPersistenceInterface
 
         $this->flushAndResetManagerOnUniqueConstraintViolation($this->entityManager);
 
+        $firstStatus = $statusCollection->first();
+        $screenName = $firstStatus instanceof StatusInterface ? $firstStatus->getScreenName() : null;
+
         return [
             'extracts'    => $propertiesCollection,
-            'screen_name' => $extract['screen_name'] ?? null,
-            'statuses'    => $statuses
+            'screen_name' => $screenName,
+            'statuses'    => $statusCollection
         ];
     }
 
@@ -142,16 +111,84 @@ class StatusPersistence implements StatusPersistenceInterface
         }
     }
 
+    private function logStatusToBeInserted(StatusInterface $status): void
+    {
+        if ($status->getId() === null) {
+            $this->collectStatusLogger->logStatus($status);
+        }
+    }
+
+    private function persistStatus(
+        CollectionInterface $statuses,
+        TaggedStatus $taggedStatus,
+        ?Aggregate $aggregate
+    ): CollectionInterface {
+        $extract = $taggedStatus->toLegacyProps();
+        $status  = $this->taggedStatusRepository
+            ->convertPropsToStatus($extract, $aggregate);
+
+        $this->logStatusToBeInserted($status);
+
+        $status = $this->unarchiveStatus($status, $this->entityManager);
+        $this->refreshUpdatedAt($status);
+
+        $this->persistTimelyStatus($aggregate, $status);
+
+        $this->entityManager->persist($status);
+
+        return $statuses->add($status);
+    }
+
+    private function persistTimelyStatus(
+        ?Aggregate $aggregate,
+        StatusInterface $status
+    ): void {
+        if ($aggregate instanceof Aggregate) {
+            $timelyStatus = $this->timelyStatusRepository->fromAggregatedStatus(
+                $status,
+                $aggregate
+            );
+            $this->entityManager->persist($timelyStatus);
+        }
+    }
+
+    private function refreshUpdatedAt(StatusInterface $status): void
+    {
+        if ($status->getId()) {
+            try {
+                $status->setUpdatedAt(
+                    new \DateTime('now', new \DateTimeZone('UTC'))
+                );
+            } catch (\Exception $exception) {
+                $this->appLogger->error($exception->getMessage());
+            }
+        }
+    }
+
+    private function tokenSetter(AccessToken $accessToken): \Closure
+    {
+        return function ($extract) use ($accessToken) {
+            $extract['identifier'] = $accessToken->accessToken();
+
+            return $extract;
+        };
+    }
+
     /**
-     * @param ArchivedStatus         $archivedStatus
+     * @param StatusInterface        $status
      * @param EntityManagerInterface $entityManager
      *
      * @return Status
      */
     private function unarchiveStatus(
-        ArchivedStatus $archivedStatus,
+        StatusInterface $status,
         EntityManagerInterface $entityManager
     ): StatusInterface {
+        if (!($status instanceof ArchivedStatus)) {
+            return $status;
+        }
+
+        $archivedStatus = $status;
         $status = Status::fromArchivedStatus($archivedStatus);
 
         $entityManager->remove($archivedStatus);
